Refactor twitter stuff into twitter_api directory
