+++ .github/workflows/python-app.yml
@@ -51,7 +51,7 @@ jobs:
     - name: Tweet
       if: matrix.os == 'ubuntu-latest' && github.ref == 'refs/heads/main' && github.event_name == 'schedule'
       run: |
-        cd bot
+        cd bot/twitter_api
         python tweet_plot.py
         git config --local user.email "41898282+github-actions[bot]@users.noreply.github.com"
         git config --local user.name "github-actions[bot]"
